judgment-reader-report-lambda/requirements.txt
psycopg2-binary
python-dotenv
pytest
pytest-cov